% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplification_curves.R
\docType{data}
\name{amplification_curves}
\alias{amplification_curves}
\title{Amplification curves}
\format{
A data frame with 24,576
amplification curves, 50 cycles each:

\describe{
\item{\code{plate}}{Plate identifier. Because one plate was used per gene, the
name of the plate is the same as the values in \code{target}.}
\item{\code{well}}{Well identifier.}
\item{\code{dye}}{In all reactions the SYBR Green I master mix (Roche) was used,
so the value is always \code{"SYBR"}.}
\item{\code{cycle}}{PCR cycle.}
\item{\code{fluor}}{Raw fluorescence values.}
}
}
\source{
\itemize{
\item Vermeulen (2009), \doi{10.1016/S1470-2045(09)70154-8}.
\item Ruijter (2013), \doi{10.1016/j.ymeth.2012.08.011}.
}
}
\usage{
amplification_curves
}
\description{
A data set containing raw fluorescence amplification data. Data is
obtained from Ruijter (2013), \doi{10.1016/j.ymeth.2012.08.011} but original
source is by Vermeulen (2009), \doi{10.1016/S1470-2045(09)70154-8}.
}
\seealso{
\code{reactions} \code{targets} \code{samples}
}
\keyword{datasets}
