% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_biomarker.R
\name{ds_biomarker}
\alias{ds_biomarker}
\title{Biomarker data set}
\source{
\itemize{
\item \url{https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip}
\item \url{https://github.com/ramiromagno/vermeulen/tree/main/data-raw}
}
}
\usage{
ds_biomarker()
}
\value{
A data frame with 24,576 amplification curves, 50 cycles each:

\describe{
\item{\code{plate}}{Plate identifier. Because one plate was used per gene, the
name of the plate is the same as the values in \code{target}.}
\item{\code{well}}{Well identifier.}
\item{\code{dye}}{In all reactions the SYBR Green I master mix (Roche) was used,
so the value is always \code{"SYBR"}.}
\item{\code{target}}{Target identifier, in almost all cases the name of a gene.}
\item{\code{target_type}}{Target type: either target of interest (\code{"toi"}) or
reference target (\code{"ref"}).}
\item{\code{sample}}{Sample identifier.}
\item{\code{sample_type}}{Sample type.}
\item{\code{copies}}{Standard copy number.}
\item{\code{dilution}}{Dilution factor. Higher number means greater dilution.}
\item{\code{cycle}}{PCR cycle.}
\item{\code{fluor}}{Raw fluorescence values.}
}
}
\description{
This function retrieves the biomarker data set, a data set containing raw
fluorescence amplification data. Data was gathered from from Ruijter et al.
(2013), \doi{10.1016/j.ymeth.2012.08.011} but original source is by Vermeulen
et al. (2009), \doi{10.1016/S1470-2045(09)70154-8}.
}
\details{
The tidy version of the data is kept at the repository of the source of
\code{{vermeulen}} package. This function fetches such data and thus requires
internet connection. It takes a few seconds to run.
}
\examples{
\dontrun{
# Takes ~ 8 sec
ds_biomarker()
}

}
