% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactions.R
\name{reactions}
\alias{reactions}
\title{Reactions}
\format{
A data frame with 24,576 reactions and 5 variables:

\describe{
\item{\code{plate}}{Plate identifier. Because one plate was used per gene, the
name of the plate is the same as the values in \code{target}.}
\item{\code{well}}{Well identifier.}
\item{\code{dye}}{In all reactions the SYBR Green I master mix (Roche) was used,
so the value is always \code{"SYBR"}.}
\item{\code{sample}}{Sample identifier.}
\item{\code{target}}{Target.}
}
}
\source{
\itemize{
\item Vermeulen (2009), \doi{10.1016/S1470-2045(09)70154-8}.
\item Ruijter (2013), \doi{10.1016/j.ymeth.2012.08.011}.
}
}
\usage{
reactions()
}
\description{
A data set containing the metadata associated with each qPCR well and
respective reaction mix. Data is obtained from Ruijter (2013),
\doi{10.1016/j.ymeth.2012.08.011} but original source is by Vermeulen (2009),
\doi{10.1016/S1470-2045(09)70154-8}.
}
\seealso{
\code{amplification_curves} \code{targets} \code{samples}
}
